+++ src/Sulu/Component/Content/Mapper/ContentMapper.php
@@ -42,7 +42,6 @@ use Sulu\Component\Content\Exception\TranslatedNodeNotFoundException;
 use Sulu\Component\Content\Extension\ExtensionInterface;
 use Sulu\Component\Content\Extension\ExtensionManagerInterface;
 use Sulu\Component\Content\Mapper\Event\ContentNodeEvent;
-use Sulu\Component\Content\Mapper\Translation\TranslatedProperty;
 use Sulu\Component\Content\Types\ResourceLocatorInterface;
 use Sulu\Component\Content\Types\Rlp\Strategy\RlpStrategyInterface;
 use Sulu\Component\DocumentManager\DocumentManager;
@@ -881,32 +880,6 @@ class ContentMapper implements ContentMapperInterface
         );
 
         return $extension->getContentData($data);
-    }
-
-    /**
-     * TODO: Move this to ResourceLocator repository>.
-     *
-     * {@inheritdoc}
-     */
-    public function restoreHistoryPath($path, $userId, $webspaceKey, $locale, $segmentKey = null)
-    {
-        $this->rlpStrategy->restoreByPath($path, $webspaceKey, $locale, $segmentKey);
-
-        $content = $this->loadByResourceLocator($path, $webspaceKey, $locale, $segmentKey);
-        $property = $content->getPropertyByTagName('sulu.rlp');
-        $property->setValue($path);
-
-        $node = $this->sessionManager->getSession()->getNodeByIdentifier($content->getUuid());
-
-        $contentType = $this->contentTypeManager->get($property->getContentTypeName());
-        $contentType->write(
-            $node,
-            new TranslatedProperty($property, $locale, $this->namespaceRegistry->getPrefix('content_localized')),
-            $userId,
-            $webspaceKey,
-            $locale,
-            $segmentKey
-        );
     }
 
     private function loadDocument($pathOrUuid, $locale, $options, $shouldExclude = true)